docs: moved provider instantiation docs snippets
