+++ apps/docs-snippets/src/guide/provider/provider.test.ts
@@ -1,4 +1,4 @@
-import { Provider, ScriptTransactionRequest, WalletUnlocked, Address } from 'fuels';
+import { Provider, ScriptTransactionRequest, Address } from 'fuels';
 import { launchTestNode } from 'fuels/test-utils';
 
 /**
@@ -6,38 +6,6 @@ import { launchTestNode } from 'fuels/test-utils';
  * @group browser
  */
 describe('Provider', () => {
-  it('base examples', async () => {
-    using launched = await launchTestNode();
-
-    const mockedProvider = await Provider.create(launched.provider.url);
-    vi.spyOn(Provider, 'create').mockResolvedValueOnce(mockedProvider);
-
-    // #region provider-definition
-    // #import { Provider, WalletUnlocked };
-
-    const FUEL_NETWORK_URL = 'http://127.0.0.1:4000/v1/graphql';
-
-    // Create the provider
-    const provider = await Provider.create(FUEL_NETWORK_URL);
-
-    // Querying the blockchain
-    const { consensusParameters } = provider.getChain();
-
-    // Create a new wallet
-    const wallet = WalletUnlocked.generate({ provider });
-
-    // Get the balances of the wallet (this will be empty until we have assets)
-    const { balances } = await wallet.getBalances();
-    // []
-    // #endregion provider-definition
-
-    expect(provider).toBeDefined();
-    expect(provider).toBeInstanceOf(Provider);
-    expect(consensusParameters).toBeDefined();
-    expect(consensusParameters).toBeInstanceOf(Object);
-    expect(balances).toEqual([]);
-  });
-
   it('fetches the base asset ID', async () => {
     const recipientAddress = Address.fromRandom();
     using launched = await launchTestNode();
